Move math to GooeyMath + Guava ranges
